remove evaluation code from grounded_mobile_sam, refactor
