+++ projects/habitat_ovmm/configs/agent/heuristic_agent.yaml
@@ -4,6 +4,7 @@ exploration_strategy: seen_frontier  # exploration strategy ("seen_frontier", "b
 radius: 0.05            # robot radius (in meters)
 fall_wait_steps: 200    # number of steps to wait after the object has been dropped
 store_all_categories: False  # whether to store all semantic categories in the map or just task-relevant ones
+detection_module: detic # the detector to use for perception in case ground_truth_semantics are turned off
 SEMANTIC_MAP:
   semantic_categories: rearrange # map semantic channel categories ("coco_indoor", "longtail_indoor", "mukul_indoor", "rearrange")
   num_sem_categories: 5    # Following 5 categories: ["misc", "object_category", "start_receptacle", "goal_receptacle", "others"]
