+++ src/home_robot/home_robot/perception/detection/coco_maskrcnn/coco_maskrcnn.py
@@ -14,6 +14,7 @@ from typing import List, Optional, Tuple
 
 import numpy as np
 import torch
+
 from detectron2.checkpoint import DetectionCheckpointer
 from detectron2.config import get_cfg
 from detectron2.data.catalog import MetadataCatalog
