+++ src/home_robot/home_robot/perception/detection/detic/detic_perception.py
@@ -13,13 +13,14 @@ from typing import List, Optional, Tuple
 import cv2
 import numpy as np
 import torch
+
 from detectron2.config import get_cfg
 from detectron2.data import MetadataCatalog
 from detectron2.engine.defaults import DefaultPredictor
 from detectron2.utils.visualizer import ColorMode, Visualizer
-
 from home_robot.core.abstract_perception import PerceptionModule
 from home_robot.core.interfaces import Observations
+from home_robot.perception.detection.utils import filter_depth, overlay_masks
 
 sys.path.insert(
     0, str(Path(__file__).resolve().parent / "Detic/third_party/CenterNet2/")
@@ -61,42 +62,6 @@ def get_clip_embeddings(vocabulary, prompt="a "):
     texts = [prompt + x for x in vocabulary]
     emb = text_encoder(texts).detach().permute(1, 0).contiguous().cpu()
     return emb
-
-
-def overlay_masks(
-    masks: np.ndarray, class_idcs: np.ndarray, shape: Tuple[int, int]
-) -> np.ndarray:
-    """Overlays the masks of objects
-    Determines the order of masks based on mask size
-    """
-    mask_sizes = [np.sum(mask) for mask in masks]
-    sorted_mask_idcs = np.argsort(mask_sizes)
-
-    semantic_mask = np.zeros(shape)
-    instance_mask = -np.ones(shape)
-    for i_mask in sorted_mask_idcs[::-1]:  # largest to smallest
-        semantic_mask[masks[i_mask].astype(bool)] = class_idcs[i_mask]
-        instance_mask[masks[i_mask].astype(bool)] = i_mask
-
-    return semantic_mask, instance_mask
-
-
-def filter_depth(
-    mask: np.ndarray, depth: np.ndarray, depth_threshold: Optional[float] = None
-) -> np.ndarray:
-    md = np.median(depth[mask == 1])  # median depth
-    if md == 0:
-        # Remove mask if more than half of points has invalid depth
-        filter_mask = np.ones_like(mask, dtype=bool)
-    elif depth_threshold is not None:
-        # Restrict objects to 1m depth
-        filter_mask = (depth >= md + depth_threshold) | (depth <= md - depth_threshold)
-    else:
-        filter_mask = np.zeros_like(mask, dtype=bool)
-    mask_out = mask.copy()
-    mask_out[filter_mask] = 0.0
-
-    return mask_out
 
 
 class DeticPerception(PerceptionModule):
@@ -242,7 +207,6 @@ class DeticPerception(PerceptionModule):
         image = cv2.cvtColor(obs.rgb, cv2.COLOR_RGB2BGR)
         depth = obs.depth
         height, width, _ = image.shape
-
         pred = self.predictor(image)
 
         if obs.task_observations is None:
